# assets/docs/swagger-api-server/projects/panda-api-docs/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0.0"
  title: Athens (Panda) API
  description: API Documentation for Athens API Server. Please be inform some of our routes may not be resposive due to other socket authentication which Swagger hasn't. Our target host can be located on **localhost:8000**. If you have further questions and clarifications you may contact us immediately.
  termsOfService: 'http://dev2.neworientalclub88.com'
  contact:
    email: jhameil@neworientalclub88.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
# during dev, should point to your local machine
host: localhost:8000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - Authorization
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
...
  /betplace:
    get:
      tags:
        - Betting
      description: Used to bet via id
      summary: Please be informed that some of our routes may not respond due to WebSocket requirements/authentication.
      operationId: bet_place
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - in: query
          name: token
          type: string
          required: true
        - in: query
          name: limit
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            properties:
              message:
                type: string
              data:
                type: string
        # responses may fall through to errors
        default:
          description: Error
          schema:
            properties:
              message:
                type: string
...
  /dealergift/getById:
    get:
      tags:
        - Dealers
      description: Used to get dealer gift via ID
      summary: Please be informed that some of our routes may not respond due to WebSocket requirements/authentication.
      operationId: dealer_gift_byId
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - in: query
          name: token
          description: Token is need to verify your credentials
          required: true
          type: string
        - in: query
          name: dealer_id
          description: Dealear ID
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            type: string
        # responses may fall through to errors
        default:
          description: Error
          schema:
            properties:
              message:
                type: string
  /dealer/getById:
    get:
      tags:
        - Dealers
      description: Used to get dealear information by ID
      summary: Please be informed that some of our routes may not respond due to WebSocket requirements/authentication.
      operationId: dealer_get_byId
      consumes:
        - application/json
      parameters:
        - in: query
          name: token
          description: Token is need to verify your credentials
          required: true
          type: string
        - in: query
          name: id
          description: Dealer ID
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            type: string
        # responses may fall through to errors
        default:
          description: Error
          schema:
            properties:
              message:
                type: string
...